file upload problme solved

+++ src/app/utils/hostImageToCloudinary.ts
@@ -1,7 +1,9 @@
 import { v2 as cloudinary, UploadApiResponse } from 'cloudinary';
 import fs from 'fs';
+import path from 'path';
 import config from '../config';
 import multer from 'multer';
+import os from 'os';
 
 // -------------Configuration values for cloudinary
 cloudinary.config({
@@ -15,6 +17,9 @@ export const hostImageToCloudinary = (
   imagePath: string,
 ): Promise<Record<string, unknown>> => {
   return new Promise((resolve, reject) => {
+    if (!fs.existsSync(imagePath)) {
+      return reject(new Error('File not found for uploading to Cloudinary'));
+    }
     // -------------Upload an image to cloudinary
     cloudinary.uploader.upload(
       imagePath,
@@ -27,13 +32,15 @@ export const hostImageToCloudinary = (
         }
         resolve(result as UploadApiResponse);
         // delete a file asynchronously
-        fs.unlink(imagePath, (err) => {
-          if (err) {
-            console.log(err);
-          } else {
-            console.log('Image hosted and Local file is deleted successfully.');
-          }
-        });
+        if (fs.existsSync(imagePath)) {
+          fs.unlink(imagePath, (err) => {
+            if (err) {
+              console.log(err);
+            } else {
+              console.log('Image hosted and Local file is deleted successfully.');
+            }
+          });
+        }
       },
     );
   });
@@ -42,7 +49,11 @@ export const hostImageToCloudinary = (
 // -------------Configuration for multer for temporary storage and file management
 const storage = multer.diskStorage({
   destination: function (req, file, cb) {
-    cb(null, process.cwd() + '/uploads/');
+    const uploadPath = path.join(os.tmpdir(), 'shraban_uploads');
+    if (!fs.existsSync(uploadPath)) {
+      fs.mkdirSync(uploadPath, { recursive: true });
+    }
+    cb(null, uploadPath);
   },
   filename: function (req, file, cb) {
     const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1e9);
