+++ src/app/utils/hostPdfToR2.ts
@@ -1,6 +1,8 @@
 import { PutObjectCommand, S3Client } from '@aws-sdk/client-s3';
 import config from '../config';
 import fs from 'fs';
+import AppError from '../errors/AppError';
+import httpStatus from 'http-status';
 
 const s3Client = new S3Client({
   region: 'auto',
@@ -16,6 +18,13 @@ export const hostPdfToR2 = async (
   filePath: string,
   contentType: string = 'application/pdf',
 ): Promise<string> => {
+  if (!fs.existsSync(filePath)) {
+    throw new AppError(
+      httpStatus.NOT_FOUND,
+      'File not found for uploading to R2',
+    );
+  }
+
   const fileBuffer = fs.readFileSync(filePath);
 
   const command = new PutObjectCommand({
@@ -31,13 +40,15 @@ export const hostPdfToR2 = async (
   const publicUrl = `${config.r2_public_url}/${fileName}`;
 
   // Delete local file after upload
-  fs.unlink(filePath, (err) => {
-    if (err) {
-      console.error('Error deleting local PDF file:', err);
-    } else {
-      console.log('PDF hosted on R2 and Local file is deleted successfully.');
-    }
-  });
+  if (fs.existsSync(filePath)) {
+    fs.unlink(filePath, (err) => {
+      if (err) {
+        console.error('Error deleting local PDF file:', err);
+      } else {
+        console.log('PDF hosted on R2 and Local file is deleted successfully.');
+      }
+    });
+  }
 
   return publicUrl;
 };
